refactor(shared): the zod helpers are library extensions, so they live in services
`createUniqueArraySchema` and `createNormalizedStringSchema` both wrap zod, and a
helper that imports a third-party package belongs in `services/` rather than
`util/` — `util/` is for utilities with no external dependency. `util/zod` is gone
and `normalizeString` is left in `util/text` alone, importing nothing.

The two types annotating `createUniqueArraySchema` move to `models/zod`, one export
per file. That closes the ledger's open finding without narrowing the published
`.d.ts`: both types stay exported, so `@esposter/shared` keeps the surface it had.
