+++ .agents/ledgers/file-organization.md
@@ -4,7 +4,7 @@ The question is where a thing lives and whether it exists twice — one export p
 
 | Unit                                                                | Swept      | Notes                                                                                                                                                                                |
 | ------------------------------------------------------------------- | ---------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| `packages/shared`, `packages/shared-node`                           | 2026-08-30 | every export has a caller; the two with none are the same file's own annotations, raised below rather than done                                                                      |
+| `packages/shared`, `packages/shared-node`                           | 2026-08-30 | the two zod helpers moved to `services/zod`, their annotating types to `models/zod`, one export per file                                                                             |
 | `app/shared/services`, `app/shared/util`                            | 2026-08-27 | `AchievementDefinitionMap` moved out of `achievementDefinitions.ts` into its own file, with its colocated suite; `getSynchronizedFunction`'s second export is exempt and now says so |
 | `app/shared/models`                                                 | —          | 319 files; splits again at its own subdirectories on contact                                                                                                                         |
 | `app/services`, `app/util`, `app/models`, `app/types`               | —          | models vs services vs utils vs constants; duplicate constants                                                                                                                        |
@@ -29,14 +29,6 @@ A duplicate constant is the one thing no skill states a grep for, because it is 
 grep -rhoE '"[a-zA-Z][a-zA-Z0-9 ./_-]{4,}"' --include=*.ts --include=*.vue packages/app/app packages/app/server packages/app/shared packages/*/src |
   sort | uniq -c | sort -rn | awk '$1 > 1'
 ```
-
-## Open findings
-
-- **`createUniqueArraySchema.ts` exports two types nothing outside it names.** `CreateUniqueArraySchema`
-  Annotates the const beside it and `UniqueArraySchemaKey` is used only by that annotation, so both are the
-  File's own scaffolding rather than a surface — the no-unused-exports rule says they lose the `export`, and
-  Building with it removed passes `attw` and `publint`. Left alone because `@esposter/shared` is published, and
-  Narrowing a published `.d.ts` is a breaking change to raise rather than fold into a sweep.
 
 ## The consumer scan
 
@@ -73,5 +65,9 @@ this referenced at all" — the second is what separates a dead export from one 
 ## Next enforceable
 
 - One export per file is syntactic; a custom oxlint plugin decides it outright.
+- A `util/` file importing a third-party package belongs in `services/`, and that is a specifier test: a
+  `no-restricted-imports` override on `**/util/**` whose `group` is `["*", "!node:*", "!#src/*", "!@esposter/*"]` decides
+  it, if oxlint honours a negated group and an `allowTypeImports` escape for the pure type utilities under
+  `util/types`. Both are unverified — the pass that builds it proves the rule can fail first (`sweeps` skill).
 - Alias imports are already enforced — the `@/**`-under-`packages/*/src/**` ban retired the `package-imports` sweep.
 - Duplicate constants and the sole-consumer rule need the whole repo in mind; they stay with the sweep.
