+++ packages/shared/src/models/zod/UniqueArraySchemaKey.ts
@@ -1,0 +1,5 @@
+import type { z } from "zod";
+
+export type UniqueArraySchemaKey<TSchema extends z.ZodType> = TSchema extends z.ZodObject
+  ? keyof TSchema["shape"] & string
+  : keyof z.output<TSchema> & string;
