+++ packages/shared/src/util/text/normalizeString.test.ts
@@ -1,4 +1,4 @@
-import { createNormalizedStringSchema, normalizeString } from "#src/util/text/normalizeString";
+import { normalizeString } from "#src/util/text/normalizeString";
 import { describe, expect, test } from "vitest";
 
 describe(normalizeString, () => {
@@ -9,14 +9,5 @@ describe(normalizeString, () => {
     expect(normalizeString(" ")).toBe("");
     expect(normalizeString(null)).toBe("");
     expect(normalizeString(undefined)).toBe("");
-  });
-});
-
-describe(createNormalizedStringSchema, () => {
-  test("trims whitespace before validating", () => {
-    expect.hasAssertions();
-
-    expect(createNormalizedStringSchema(5).parse(" a ")).toBe("a");
-    expect(createNormalizedStringSchema(5).parse(" ")).toBe("");
   });
 });
